Fix possible bug
We checked whether showMyLocation was a non-false boolean (always the
case if the property is not set) and not whether the property was really
set. This will work when the default is "false" and we set it to "true",
but will fail the other way around. Some other use cases with this
implementation may also lead to unexpected results.

Change-Id: I59fe5e7aee4349d084edf434434507c38579bf0c