Fix example page titles on iOS
Static page class getter 'name' was colliding with the function property
'name'. On Android, the static class getter overwrote the function
property, but on iOS it didn't, which resulted in the class constructor
name instead of the given page name being displayed.

Use capital case for the property to designate it as static and thus
prevent collision with the function property 'name'.

Change-Id: I1943e9e32f6dfd2bfbf2eb28fb45adff6e8d8551