Fix app crash in Tabris 3.5 version
The plugin crashes the app when we use it with the Tabris 3.5 version on Android. The issue occurs due to SupportMapFragment which is added before MapHolderView was attached to the parent.

The change solves the problem and also updates the configuration file to make the project compatible with Tabris 3.x.