Upgrade Google Maps SDK (#75)
Google Maps Platform has introduced a new renderer for Maps SDK for
Android version 18.0, initially available on an opt-in basis. Starting
from version 18.2, it will be the default renderer, with an option to
specify the Legacy renderer (Renderer.LEGACY).

Apps built with the Maps SDK for Android will be automatically updated
to the latest renderer beginning March 4, 2024. The legacy renderer is
scheduled for decommissioning on March 3, 2025.

Update the Google Maps SDK dependency to the latest version for testing
the application with the new renderer before the scheduled automatic
update in March 2024, in order to limit risks of potential issues.

Additionally, update the Gradle versions and build files to align with
the configuration used by the Tabris Android platform.